Move ws functions to utils
